// client/src/context/actions.ts
export enum WordsActions {
  SET_WORDS_LIST = 'SET_WORDS_LIST',
  SET_CURRENT_WORD = 'SET_CURRENT_WORD',
  SET_IS_CORRECT = 'SET_IS_CORRECT',
  SET_IS_DISABLED = 'SET_IS_DISABLED',
  SET_IS_FINISHED = 'SET_IS_FINISHED',
  SET_SCORE = 'SET_SCORE',
  SET_COUNT = 'SET_COUNT',
  SET_INITIAL_STATE = 'SET_INITIAL_STATE'
}